+++ utils/stories/src/AccessibilityAudit/constants.tsx
@@ -1,19 +1,13 @@
 import { SettingsOutlineIcon, ShieldCheckOutlineIcon, ProgressCheckIcon } from '@ultraviolet/icons'
 import { Text } from '@ultraviolet/ui'
-import type { ReactNode } from 'react'
+import type { A11yLevel, A11yLevelInfo } from './types'
 
-export type A11yLevel = 'partial' | 'compliant' | 'certified'
-
-export type A11yLevelInfo = {
-  level: A11yLevel
-  label: string
-  icon: ReactNode
-  description: ReactNode
-}
-
-export const WCAG_PRINCIPLES = ['perceivable', 'operable', 'understandable', 'robust'] as const
-
-export type WcagPrinciple = (typeof WCAG_PRINCIPLES)[number]
+export const WCAG_PRINCIPLES = [
+  { name: 'perceivable', label: 'Perceivable' },
+  { name: 'operable', label: 'Operable' },
+  { name: 'understandable', label: 'Understandable' },
+  { name: 'robust', label: 'Robust' },
+] as const
 
 export const A11Y_LEVELS: Record<A11yLevel, A11yLevelInfo> = {
   partial: {
@@ -49,42 +43,4 @@ export const A11Y_LEVELS: Record<A11yLevel, A11yLevelInfo> = {
       </Text>
     ),
   },
-}
-
-export type ComponentA11yStatus = Record<WcagPrinciple, boolean>
-
-export type AuditCategories = {
-  id: string
-  completed: boolean
-}[]
-
-export type ComponentStoryParameters = {
-  deprecated?: boolean
-  experimental?: boolean
-  a11yStatus?: ComponentA11yStatus
-}
-
-export const getComponentA11yStatus = (parameters: ComponentStoryParameters): ComponentA11yStatus => {
-  if (!parameters?.a11yStatus) {
-    return Object.fromEntries(WCAG_PRINCIPLES.map(principle => [principle, false])) as ComponentA11yStatus
-  }
-
-  return parameters.a11yStatus
-}
-
-export const getComponentAuditCategories = (parameters: ComponentStoryParameters): AuditCategories => {
-  const a11yStatus = getComponentA11yStatus(parameters)
-
-  return WCAG_PRINCIPLES.map(principle => ({
-    id: principle,
-    completed: a11yStatus[principle] ?? false,
-  }))
-}
-
-export const findA11yLevel = (parameters: { a11yStatus?: ComponentA11yStatus }): A11yLevel => {
-  if (!parameters?.a11yStatus) {
-    return 'partial'
-  }
-
-  return WCAG_PRINCIPLES.every(principle => parameters.a11yStatus?.[principle] === true) ? 'compliant' : 'partial'
 }
